# backend/artifacts/openapi.json
{
  "components": {
    "schemas": {
      "DependencyStatus": {
        "properties": {
          "name": {
            "examples": [
              "database"
            ],
            "title": "Name",
            "type": "string"
          },
          "status": {
            "enum": [
              "ok",
              "unavailable"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "name",
          "status"
        ],
        "title": "DependencyStatus",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "LivenessResponse": {
        "properties": {
          "status": {
            "const": "live",
            "default": "live",
            "title": "Status",
            "type": "string"
          }
        },
        "title": "LivenessResponse",
        "type": "object"
      },
      "LoginRequest": {
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
      "LogoutResponse": {
        "properties": {
          "revoked": {
            "description": "Whether a live session was revoked by this call.",
            "title": "Revoked",
            "type": "boolean"
          }
        },
        "required": [
          "revoked"
        ],
        "title": "LogoutResponse",
        "type": "object"
      },
      "ProblemDetail": {
        "description": "RFC 7807 problem document, extended with `code` and `request_id`.\n\n`code` is the stable, machine-readable identifier the frontend switches on; `title` and\n`detail` are human-facing text that may be reworded without breaking clients.",
        "properties": {
          "code": {
            "examples": [
              "RESOURCE_NOT_FOUND"
            ],
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "examples": [
              "The requested tender does not exist."
            ],
            "title": "Detail",
            "type": "string"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Field-level detail for validation failures.",
            "title": "Errors"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "examples": [
              "/api/v1/tenders/8f0c.../documents"
            ],
            "title": "Instance"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "status": {
            "examples": [
              404
            ],
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "examples": [
              "Resource not found"
            ],
            "title": "Title",
            "type": "string"
          },
          "type": {
            "examples": [
              "https://bidpilot.dev/problems/not-found"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "code"
        ],
        "title": "ProblemDetail",
        "type": "object"
      },
      "ReadinessResponse": {
        "properties": {
          "dependencies": {
            "items": {
              "$ref": "#/components/schemas/DependencyStatus"
            },
            "title": "Dependencies",
            "type": "array"
          },
          "status": {
            "enum": [
              "ready",
              "degraded"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "dependencies"
        ],
        "title": "ReadinessResponse",
        "type": "object"
      },
      "RegisterRequest": {
        "properties": {
          "display_name": {
            "examples": [
              "Tender Coordinator"
            ],
            "maxLength": 120,
            "minLength": 1,
            "title": "Display Name",
            "type": "string"
          },
          "email": {
            "examples": [
              "coordinator@fm-demo.ae"
            ],
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "examples": [
              "a-long-passphrase-1"
            ],
            "maxLength": 128,
            "minLength": 12,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password",
          "display_name"
        ],
        "title": "RegisterRequest",
        "type": "object"
      },
      "SessionRead": {
        "description": "A live refresh session, so a user can see where they are signed in.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "user_agent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Agent"
          }
        },
        "required": [
          "id",
          "created_at",
          "expires_at",
          "user_agent"
        ],
        "title": "SessionRead",
        "type": "object"
      },
      "TokenResponse": {
        "description": "Access token plus the authenticated user.\n\nThe refresh token is absent by design: it is delivered only as an HttpOnly cookie, so\nJavaScript cannot read it and an XSS payload cannot exfiltrate a long-lived credential.",
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "expires_in": {
            "description": "Access token lifetime in seconds.",
            "title": "Expires In",
            "type": "integer"
          },
          "token_type": {
            "const": "bearer",
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/UserRead"
          }
        },
        "required": [
          "access_token",
          "expires_in",
          "user"
        ],
        "title": "TokenResponse",
        "type": "object"
      },
      "UserRead": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "email",
          "display_name",
          "is_active",
          "created_at"
        ],
        "title": "UserRead",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "Access token from /auth/login",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "BidPilot UAE turns an unstructured tender PDF into a reviewable decision workspace:\npage-aware extraction, schema-constrained requirement extraction, verified citations,\ndeterministic readiness scoring, and human review.\n\nUploaded document content is treated as untrusted evidence, never as instructions.\nAI output is advisory; the readiness score is calculated in Python, not by a model.",
    "title": "BidPilot UAE",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "description": "Returns a short-lived access token and sets an HttpOnly refresh cookie. A failed attempt always returns the same message, whether or not the account exists.",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Too Many Requests"
          }
        },
        "summary": "Sign in with email and password",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "description": "Idempotent: clears the cookie and returns 200 whether or not the presented token was still live, so the endpoint cannot be used to probe token validity.",
        "operationId": "logout_api_v1_auth_logout_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogoutResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Revoke the current session",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/v1/auth/logout-all": {
      "post": {
        "operationId": "logout_all_api_v1_auth_logout_all_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogoutResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Unauthorized"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Too Many Requests"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Revoke every session for the signed-in user",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "operationId": "me_api_v1_auth_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Unauthorized"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Too Many Requests"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "The signed-in user",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "description": "Rotates the refresh token: the presented one is revoked and a new cookie is set. Presenting an already-revoked token revokes every session for that user.",
        "operationId": "refresh_api_v1_auth_refresh_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Unauthorized"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Too Many Requests"
          }
        },
        "summary": "Exchange the refresh cookie for a new access token",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "operationId": "register_api_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Too Many Requests"
          }
        },
        "summary": "Create an account and sign in",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/v1/auth/sessions": {
      "get": {
        "description": "Where this account is currently signed in. Revoked and expired sessions are omitted.",
        "operationId": "list_sessions_api_v1_auth_sessions_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SessionRead"
                  },
                  "title": "Response List Sessions Api V1 Auth Sessions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Unauthorized"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Too Many Requests"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Live sessions for the signed-in user",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/v1/health/live": {
      "get": {
        "description": "Returns 200 whenever the process can serve requests. Checks no dependencies.",
        "operationId": "liveness_api_v1_health_live_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LivenessResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Liveness probe",
        "tags": [
          "health"
        ]
      }
    },
    "/api/v1/health/ready": {
      "get": {
        "description": "Verifies PostgreSQL and Redis connectivity. Returns 503 problem+json naming the unavailable dependency when either check fails.",
        "operationId": "readiness_api_v1_health_ready_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadinessResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Readiness probe",
        "tags": [
          "health"
        ]
      }
    }
  }
}